keepFiles

+++ packages/css-to-tailwind/src/validate.mjs
@@ -62,7 +62,7 @@ function getSelector(css) {
 
 const LOREM_IPSUM = `Lorem ipsum dolor sit amet, consectetur adipiscing elit. Etiam nec feugiat arcu. Pellentesque sed nunc vehicula, egestas erat sodales, iaculis elit. In placerat laoreet mauris non volutpat.`;
 
-export async function validate(css, utilities) {
+export async function validate({ css, utilities, keepFiles = false }) {
   const cssSelector = getSelector(css);
   const cssFromUtilities = await utilitiesToCSS(utilities);
 
@@ -108,38 +108,45 @@ export async function validate(css, utilities) {
   await browser.close();
 
   if (!bufferA.equals(bufferB)) {
-    const location = path.resolve(__dirname, `../screenshots`);
-    const filePrefix = getFilePrefix();
+    if (keepFiles) {
+      const location = path.resolve(__dirname, `../screenshots`);
+      const filePrefix = getFilePrefix();
 
-    await fs.writeFile(`${location}/${filePrefix}_from_css.png`, bufferA);
-    await fs.writeFile(`${location}/${filePrefix}_from_utilities.png`, bufferB);
-
-    try {
-      // check if "magick" is available
-      childProcess.execSync('magick --version', { stdio: 'ignore' });
-    } catch (err) {
-      console.warn(
-        'ImageMagick is not available, skipping diff image generation',
-      );
-    }
-
-    try {
-      const command = [
-        'magick compare',
-        '-verbose',
-        '-metric mae',
-        `${location}/${filePrefix}_from_css.png`,
+      await fs.writeFile(`${location}/${filePrefix}_from_css.png`, bufferA);
+      await fs.writeFile(
         `${location}/${filePrefix}_from_utilities.png`,
-        `${location}/${filePrefix}_diff.png`,
-      ];
+        bufferB,
+      );
 
-      childProcess.execSync(command.join(' '), {
-        // do not throw error if command fails
-        stdio: 'ignore',
-      });
+      try {
+        // check if "magick" is available
+        childProcess.execSync('magick --version', { stdio: 'ignore' });
+      } catch (err) {
+        console.warn(
+          'ImageMagick is not available, skipping diff image generation',
+        );
+      }
 
-      console.log(`Diff image generated at ${location}/${filePrefix}_diff.png`);
-    } catch (err) {}
+      try {
+        const command = [
+          'magick compare',
+          '-verbose',
+          '-metric mae',
+          `${location}/${filePrefix}_from_css.png`,
+          `${location}/${filePrefix}_from_utilities.png`,
+          `${location}/${filePrefix}_diff.png`,
+        ];
+
+        childProcess.execSync(command.join(' '), {
+          // do not throw error if command fails
+          stdio: 'ignore',
+        });
+
+        console.log(
+          `Diff image generated at ${location}/${filePrefix}_diff.png`,
+        );
+      } catch (err) {}
+    }
 
     throw new Error('CSS and utilities do not match');
   }
